files-reg: temporary remount writable the mount we do unlink on
Previousely I din't mention this case because we had bad error handling
in ghost cleanup path.

Without these patch but with proper error handling for unlink we have an
error in mntn_ghost01 test:

Error (criu/files-reg.c:2269): Failed to unlink the remap file:
Read-only file system

https://jira.sw.ru/browse/PSBM-101433

Fixes: fd0a3cd9efb9 ("mount: remount ro mounts writable before
ghost-file restore")

+++ criu/files-reg.c
@@ -2276,6 +2276,11 @@ ext:
 
 	if (rfi->remap) {
 		if (!rfi->remap->is_dir) {
+			struct mount_info *mi = lookup_mnt_id(rfi->rfe->mnt_id);
+
+			if (try_remount_writable(mi, true))
+				goto err;
+
 			pr_debug("%d: Unlink: %s\n", rfi->rfe->mnt_id, rfi->path);
 			if (unlinkat(mntns_root, rfi->path, 0)) {
 				pr_perror("Failed to unlink the remap file");
